Further endpoint slice code
